+++ test/integration/integration.test.ts
@@ -151,6 +151,14 @@ describe('E2E integration tests', () => {
     );
   });
 
+  it('getNftsForContract with limit', async () => {
+    const nftsForNftContract = await alchemy.nft.getNftsForContract(
+      contractAddress,
+      { limit: 10 }
+    );
+    expect(nftsForNftContract.nfts.length).toEqual(10);
+  });
+
   it('getIterator', async () => {
     jest.setTimeout(15000);
     console.log('lets paginate');